requirements.txt
flask
#cowsay
gunicorn
#emoji
PyOpenSSL
